Update mainmenu.json
Partnership with Pebblehost ceased, as Pebblehost removed all perks for myself and TCP players quite a while back.

The only relationship TCP had with Pebblehost at the time of termination, was the ability to use their Affiliate system. TCP has supported Pebblehost with free advertisement and links for too long with nothing to show for it.

I am open to new partnerships.